feat: test-ng: everythings a plugin
