+++ tests-ng/Makefile
@@ -13,7 +13,7 @@ clean:
 .build:
 	mkdir .build
 
-.build/dist.tar.gz: util/build_dist.sh .build/runtime-x86_64.tar.gz .build/runtime-aarch64.tar.gz $(wildcard *.py) | .build
+.build/dist.tar.gz: util/build_dist.sh .build/runtime-x86_64.tar.gz .build/runtime-aarch64.tar.gz conftest.py $(wildcard plugins/*.py) $(wildcard test_*.py) | .build
 	./$< $@ x86_64:$(word 2,$^) aarch64:$(word 3,$^)
 
 .build/runtime-%.tar.gz: util/build_runtime.sh util/requirements.txt | .build
