+++ tests-ng/plugins/shell.py
